Revert name change for Response class
