+++ src/main/java/com/k3ntako/HTTPServer/ControllerInterface.java
@@ -1,0 +1,5 @@
+package com.k3ntako.HTTPServer;
+
+public interface ControllerInterface {
+  Response getResponse(RequestInterface request);
+}
