--- src/main/java/com/k3ntako/HTTPServer/RouteInterface.java
@@ -1,5 +1,0 @@
-package com.k3ntako.HTTPServer;
-
-public interface RouteInterface {
-  ResponseCreator getResponse(RequestInterface request);
-}
