+++ src/main/java/com/k3ntako/HTTPServer/RouteRegistry.java
@@ -3,13 +3,13 @@ package com.k3ntako.HTTPServer;
 import java.util.HashMap;
 
 public class RouteRegistry {
-  private HashMap<String, RouteInterface> getRoutes = new HashMap<>();
+  private HashMap<String, ControllerInterface> getRoutes = new HashMap<>();
 
-  public void registerGet(String url, RouteInterface route) {
+  public void registerGet(String url, ControllerInterface route) {
     getRoutes.put(url, route);
   }
 
-  public RouteInterface get(String url){
+  public ControllerInterface get(String url){
     return getRoutes.get(url);
   }
 }
