--- src/main/java/com/k3ntako/HTTPServer/routes/SimpleGet.java
@@ -1,11 +1,0 @@
-package com.k3ntako.HTTPServer.routes;
-
-import com.k3ntako.HTTPServer.RequestInterface;
-import com.k3ntako.HTTPServer.ResponseCreator;
-import com.k3ntako.HTTPServer.RouteInterface;
-
-public class SimpleGet implements RouteInterface {
-  public ResponseCreator getResponse(RequestInterface request) {
-    return new ResponseCreator();
-  }
-}
