+++ src/test/java/com/k3ntako/HTTPServer/RouteRegistrarTest.java
@@ -1,7 +1,7 @@
 package com.k3ntako.HTTPServer;
 
-import com.k3ntako.HTTPServer.routes.SimpleGet;
-import com.k3ntako.HTTPServer.routes.SimpleGetWithBody;
+import com.k3ntako.HTTPServer.controllers.SimpleGet;
+import com.k3ntako.HTTPServer.controllers.SimpleGetWithBody;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
