+++ src/test/java/com/k3ntako/HTTPServer/controllers/NotFoundTest.java
@@ -1,4 +1,4 @@
-package com.k3ntako.HTTPServer.routes;
+package com.k3ntako.HTTPServer.controllers;
 
 import com.k3ntako.HTTPServer.Request;
 import com.k3ntako.HTTPServer.mocks.ServerIOMock;
