+++ src/test/java/com/k3ntako/HTTPServer/controllers/SimpleGetTest.java
@@ -1,4 +1,4 @@
-package com.k3ntako.HTTPServer.routes;
+package com.k3ntako.HTTPServer.controllers;
 
 import com.k3ntako.HTTPServer.mocks.RequestMock;
 import org.junit.jupiter.api.Test;
@@ -7,19 +7,18 @@ import java.util.HashMap;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class SimpleGetWithBodyTest {
+class SimpleGetTest {
 
   @Test
   void getResponse() {
-    var request = new RequestMock("GET", "/simple_get_with_body", "HTTP/1.1", new HashMap<>(), "");
+    var request = new RequestMock("GET", "/simple_get", "HTTP/1.1", new HashMap<>(), "");
 
-    var simpleGetWithBody = new SimpleGetWithBody();
-    var response = simpleGetWithBody.getResponse(request);
+    var simpleGet = new SimpleGet();
+    var response = simpleGet.getResponse(request);
 
     var responseStr = response.createResponse();
     var expectedResponse = "HTTP/1.1 200 OK\r\n" +
-            "Content-Length: 11\r\n\r\n" +
-            "Hello world";
+            "Content-Length: 0\r\n\r\n";
     assertEquals(expectedResponse, responseStr);
   }
 }
